Add files via upload
This is the first public version I'm making available.  Be sure you examine the directory structure I've used here, or edit it as needed.  This initial version is not without issues WRT saving host profiles and loading them.  this is my first attempt at using GitHub, so this is more about me learning at this point, please bear with me.