Add files via upload
Values load to slider when attribute selected, and they no longer default to 50% when you mouse over the slider with an attribute selected.  Beginning of the "Create Host" function - when typing in a host name, if no suggestion, it asks if you want to create it.  Link is just a placeholder for now, it does nothing.